#216 Split resolution.rs and tick.rs into responsibility-focused files
resolution.rs becomes a thin module root over resolution/{effect,innate,ability}.rs,
separating the low-level effect-resolution engine from over-time gating and ability-cast
resolution. tick.rs's message-building code (tick message assembly, battle-update
construction, and per-player broadcast) moves into message/{tick_message,broadcast}.rs,
leaving tick.rs focused on phase dispatch only.
