+++ src/game/engagement/battle/message.rs
@@ -1,3 +1,6 @@
+pub mod broadcast;
+pub mod tick_message;
+
 use std::fmt;
 
 use serde::{Deserialize, Serialize};
